LayoutTests/imported/w3c/web-platform-tests/css/css-nesting/parsing-expected.txt


PASS .foo {
  & { color: green; }
}
PASS .foo { //color: red; color: green; }
PASS .foo {
  &.bar { color: green; }
}
PASS .foo {
  & .bar { color: green; }
}
PASS .foo {
  & > .bar { color: green; }
}
PASS .foo {
  .test > & .bar { color: green; }
}
PASS .foo {
  + .bar, .foo, > .lol { color: green; }
}
PASS .foo {
  &:is(.bar, &.baz) { color: green; }
}
PASS .foo {
  .bar& { color: green; }
}
PASS .foo {
  .bar & { color: green; }
}
PASS .foo {
  .bar > & { color: green; }
}
PASS .foo, .bar {
  & + .baz, &.qux { color: green; }
}
PASS .foo {
  & .bar & .baz & .qux { color: green; }
}
FAIL .foo {
  @media (min-width: 50px) {
  & { color: green; }
}
} assert_equals: expected ".foo {\n  @media (min-width: 50px) { color: green; }\n}" but got ".foo { }"
FAIL .foo {
  @media (min-width: 50px) { color: green; }
} assert_equals: expected ".foo {\n  @media (min-width: 50px) { color: green; }\n}" but got ".foo { }"
PASS main {
  & > section, & > article {
  & > header { color: green; }
}
}
PASS .foo {
  :has(div) { color: green; }
}
PASS .foo {
  .bar {
  && { color: green; }
}
}
PASS .foo {
 color: red; ident { color: green; }
}
PASS .foo {
 //color: comment; & { color: green; }
}
PASS .foo {
 .bar {
  functionalnotation(div) { color: green; }

}
